Test suite cleanup
